Unused variables, undeclared functions and a few other things corrected
Monotone-Parent: bd34993b7eba4f200ebb0f7bb90178a80282a54d
Monotone-Revision: e8e52aaea0f7cb226f0e496e8f7d2d7e6a2f7566

+++ events.c
@@ -4365,7 +4365,8 @@ static void dumpevent (XEvent *e)
       case ClientMessage:  name = "ClientMessage"; break;
       case MappingNotify:  name = "MappingNotify"; break;
     }
-    fprintf (tracefile, "event:  %s in window 0x%x\n", name, e->xany.window);
+    fprintf (tracefile, "event:  %s in window 0x%x\n", name,
+	     (unsigned int)e->xany.window);
     switch (e->type) {
       case KeyPress:
       case KeyRelease:
